Physics simulator works, keyboard doesn't.
I finished the very basics of the physics simulator.  The game can now
properly simulate the player's starting system.  None of the keyboard
buttons work yet.  Also, I skipped the menus used to build the character
because the buttons won't show up properly in this build.  I don't know
how to fix that.